+++ data/src/history/metadata.rs
@@ -7,8 +7,8 @@ use serde::{Deserialize, Serialize};
 use tokio::fs;
 
 use crate::history::{dir_path, Error, Kind};
-use crate::message::source;
-use crate::{isupport, Message};
+use crate::message::{source, MessageReferences};
+use crate::Message;
 
 #[derive(Debug, Clone, Default, Deserialize, Serialize)]
 pub struct Metadata {
@@ -55,54 +55,6 @@ impl FromStr for ReadMarker {
 impl fmt::Display for ReadMarker {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         self.0.to_rfc3339_opts(SecondsFormat::Millis, true).fmt(f)
-    }
-}
-
-#[derive(Debug, Clone, Default, Deserialize, Serialize)]
-pub struct MessageReferences {
-    pub timestamp: DateTime<Utc>,
-    pub id: Option<String>,
-}
-
-impl MessageReferences {
-    pub fn message_reference(
-        &self,
-        message_reference_types: &[isupport::MessageReferenceType],
-    ) -> isupport::MessageReference {
-        for message_reference_type in message_reference_types {
-            match message_reference_type {
-                isupport::MessageReferenceType::MessageId => {
-                    if let Some(id) = &self.id {
-                        return isupport::MessageReference::MessageId(id.clone());
-                    }
-                }
-                isupport::MessageReferenceType::Timestamp => {
-                    return isupport::MessageReference::Timestamp(self.timestamp);
-                }
-            }
-        }
-
-        isupport::MessageReference::None
-    }
-}
-
-impl PartialEq for MessageReferences {
-    fn eq(&self, other: &Self) -> bool {
-        self.timestamp.eq(&other.timestamp)
-    }
-}
-
-impl Eq for MessageReferences {}
-
-impl Ord for MessageReferences {
-    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
-        self.timestamp.cmp(&other.timestamp)
-    }
-}
-
-impl PartialOrd for MessageReferences {
-    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        Some(self.cmp(other))
     }
 }
