+++ data/src/message.rs
@@ -16,10 +16,9 @@ pub use self::formatting::Formatting;
 pub use self::source::Source;
 
 use crate::config::buffer::UsernameFormat;
-use crate::history::MessageReferences;
 use crate::time::{self, Posix};
 use crate::user::{Nick, NickRef};
-use crate::{ctcp, Config, Server, User};
+use crate::{ctcp, isupport, Config, Server, User};
 
 // References:
 // - https://datatracker.ietf.org/doc/html/rfc1738#section-5
@@ -1256,6 +1255,54 @@ where
     let intermediate = serde_json::Value::deserialize(deserializer)?;
 
     Ok(Option::<T>::deserialize(intermediate).unwrap_or_default())
+}
+
+#[derive(Debug, Clone, Default, Deserialize, Serialize)]
+pub struct MessageReferences {
+    pub timestamp: DateTime<Utc>,
+    pub id: Option<String>,
+}
+
+impl MessageReferences {
+    pub fn message_reference(
+        &self,
+        message_reference_types: &[isupport::MessageReferenceType],
+    ) -> isupport::MessageReference {
+        for message_reference_type in message_reference_types {
+            match message_reference_type {
+                isupport::MessageReferenceType::MessageId => {
+                    if let Some(id) = &self.id {
+                        return isupport::MessageReference::MessageId(id.clone());
+                    }
+                }
+                isupport::MessageReferenceType::Timestamp => {
+                    return isupport::MessageReference::Timestamp(self.timestamp);
+                }
+            }
+        }
+
+        isupport::MessageReference::None
+    }
+}
+
+impl PartialEq for MessageReferences {
+    fn eq(&self, other: &Self) -> bool {
+        self.timestamp.eq(&other.timestamp)
+    }
+}
+
+impl Eq for MessageReferences {}
+
+impl Ord for MessageReferences {
+    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
+        self.timestamp.cmp(&other.timestamp)
+    }
+}
+
+impl PartialOrd for MessageReferences {
+    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
+        Some(self.cmp(other))
+    }
 }
 
 #[cfg(test)]
